backend/pyproject.toml
[tool.poetry]
name = "codesprint"
version = "0.1.0"
description = "projeto final backend2"
authors = ["Leandro Oliveira <leandro.gooli@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = {extras = ["standard"], version = ">=0.34.2,<0.35.0"}
psycopg2-binary = "^2.9.9"
sqlalchemy = ">=2.0.40,<3.0.0"
alembic = ">=1.15.2,<2.0.0"
pydantic = ">=2.11.3,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
python-jose = {extras = ["cryptography"], version = ">=3.4.0,<4.0.0"}
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
